pyproject.toml
[tool.poetry]
name = "thumbnail-api"
version = "0.1.0"
packages = [{include="*", from="src"}]
description = "Long-running job API for creating 100x100 thumbnails from uploaded images"
authors = ["Josh Burns <joshburns@JoshnoiMac.local>"]
readme = "README.md"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "acceptance: marks tests as acceptance tests (deselect with '-m \"not acceptance\"')",
    "e2e: marks tests as end-to-end tests (deselect with '-m \"not e2e\"')"
]
addopts = "--cov=app --cov-report html"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
files = ["**/*.py"]
strict = true
plugins = ["pydantic.mypy", "pydantic.v1.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff.lint]
extend-select = ["I"]
select = ["E501", "I"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72

[tool.poetry.dependencies]
python = "^3.12.6"
annotated-types = "0.7.0"
anyio = "4.4.0"
click = "8.1.7"
fastapi = {extras = ["standard"], version = "^0.114.2"}
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.8"
pydantic = "2.9.1"
pydantic-core = "2.23.3"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
sniffio = "1.3.1"
starlette = "0.38.5"
typing-extensions = "4.12.2"
uvicorn = "0.30.6"
uvloop = "0.20.0"
watchfiles = "0.24.0"
websockets = "13.0.1"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.11.2"
ruff = "^0.6.5"
requests = "^2.32.3"
types-requests = "^2.32.0.20240914"
docker = "^7.1.0"
types-docker = "^7.1.0.20240827"
pytest-cov = "^5.0.0"
types-pyyaml = "^6.0.12.20240917"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
